Moved Constants to __init__ of utils. Added dragonfly experiments to experiment_settings.json.
