+++ HPOlibExperimentUtils/utils/__init__.py
@@ -1,0 +1,4 @@
+class Constants:
+    fixed_benchmark_settings = ['fidelity_name', 'fidelity_type', 'import_from', 'import_benchmark']
+    smac_traj_filename = "traj.json"
+    common_trajectory_filename = "traj_hpolib.json"
